+++ airlock/file_browser_api.py
@@ -1,0 +1,81 @@
+import pathlib
+from dataclasses import dataclass
+
+from airlock.api import AirlockContainer
+
+
+@dataclass(frozen=True)
+class PathItem:
+    """
+    This provides a thin abstraction over `pathlib.Path` objects with two goals:
+
+        1. Paths should be enforced as being relative to a certain "container" directory
+           and it should not be possible to traverse outside of this directory or to
+           construct one which points outside this directory (using the designated
+           constructor classmethods).
+
+        2. The abstraction should permit us, in future, to switch the implementation to
+           something which is not tied to concrete filesystem paths.
+    """
+
+    container: AirlockContainer
+    relpath: pathlib.Path
+
+    def __post_init__(self):
+        # ensure relpath is a Path
+        object.__setattr__(self, "relpath", pathlib.Path(self.relpath))
+        # ensure path is within container
+        self._absolute_path().resolve().relative_to(self.container.root())
+
+    def _absolute_path(self):
+        return self.container.root() / self.relpath
+
+    def exists(self):
+        return self._absolute_path().exists()
+
+    def is_directory(self):
+        return self._absolute_path().is_dir()
+
+    def name(self):
+        return self.relpath.name
+
+    def url(self):
+        suffix = "/" if self.is_directory() else ""
+        return self.container.get_url_for_path(f"{self.relpath}{suffix}")
+
+    def parent(self):
+        if self.relpath.parents:
+            return PathItem(self.container, self.relpath.parent)
+
+    def children(self):
+        root = self.container.root()
+        return [
+            PathItem(self.container, child.relative_to(root))
+            for child in self._absolute_path().iterdir()
+        ]
+
+    def siblings(self):
+        if not self.relpath.parents:
+            return []
+        else:
+            return self.parent().children()
+
+    def contents(self):
+        return self._absolute_path().read_text()
+
+    @property
+    def suffix(self):
+        return self.relpath.suffix
+
+    def breadcrumbs(self):
+        item = self
+        crumbs = [item]
+
+        parent = item.parent()
+        while parent:
+            if parent.relpath != pathlib.Path("."):
+                crumbs.append(parent)
+            parent = parent.parent()
+
+        crumbs.reverse()
+        return crumbs
